requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
absl-py==2.3.0
    # via tensorboard
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.13
    # via
    #   -r requirements.in
    #   fsspec
aioice==0.10.1
    # via aiortc
aiortc==1.13.0
    # via -r requirements.in
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyascii==0.3.2
    # via coqui-tts
anyio==4.9.0
    # via httpx
attrs==25.3.0
    # via aiohttp
audioread==3.0.1
    # via librosa
av==14.4.0
    # via
    #   -r requirements.in
    #   aiortc
    #   faster-whisper
babel==2.17.0
    # via gruut
bidict==0.23.1
    # via python-socketio
blinker==1.9.0
    # via flask
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via
    #   aiortc
    #   cryptography
    #   pylibsrtp
    #   sounddevice
    #   soundfile
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   duckduckgo-search
    #   flask
colorama==0.4.6
    # via
    #   click
    #   tqdm
coloredlogs==15.0.1
    # via onnxruntime
contourpy==1.3.2
    # via matplotlib
coqpit-config==0.2.0
    # via
    #   coqui-tts
    #   coqui-tts-trainer
coqui-tts==0.26.2
    # via -r requirements.in
coqui-tts-trainer==0.2.3
    # via coqui-tts
cryptography==45.0.4
    # via
    #   aiortc
    #   pyopenssl
ctranslate2==4.6.0
    # via faster-whisper
cycler==0.12.1
    # via matplotlib
cython==3.1.2
    # via coqui-tts
dateparser==1.1.8
    # via gruut
decorator==5.2.1
    # via librosa
dlib==20.0.0
    # via -r requirements.in
dnspython==2.7.0
    # via aioice
docopt==0.6.2
    # via num2words
duckdb==1.3.1
    # via -r requirements.in
duckduckgo-search==8.0.4
    # via -r requirements.in
einops==0.8.1
    # via
    #   coqui-tts
    #   encodec
encodec==0.1.1
    # via coqui-tts
faster-whisper==1.1.1
    # via -r requirements.in
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
flask==3.1.1
    # via
    #   -r requirements.in
    #   flask-socketio
flask-socketio==5.5.1
    # via -r requirements.in
flatbuffers==25.2.10
    # via onnxruntime
fonttools==4.58.4
    # via matplotlib
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2025.5.1
    # via
    #   coqui-tts
    #   coqui-tts-trainer
    #   huggingface-hub
    #   torch
google-crc32c==1.7.1
    # via aiortc
grpcio==1.73.0
    # via tensorboard
gruut[de,es,fr]==2.4.0
    # via coqui-tts
gruut-ipa==0.13.0
    # via gruut
gruut-lang-de==2.0.1
    # via gruut
gruut-lang-en==2.0.1
    # via gruut
gruut-lang-es==2.0.1
    # via gruut
gruut-lang-fr==2.0.2
    # via gruut
h11==0.16.0
    # via
    #   httpcore
    #   wsproto
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via ollama
huggingface-hub==0.33.0
    # via
    #   faster-whisper
    #   tokenizers
    #   transformers
humanfriendly==10.0
    # via coloredlogs
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
ifaddr==0.2.0
    # via aioice
inflect==7.5.0
    # via coqui-tts
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via
    #   flask
    #   torch
joblib==1.5.1
    # via
    #   librosa
    #   scikit-learn
jsonlines==1.2.0
    # via gruut
kiwisolver==1.4.8
    # via matplotlib
lazy-loader==0.4
    # via librosa
librosa==0.11.0
    # via coqui-tts
llvmlite==0.44.0
    # via numba
lxml==5.4.0
    # via duckduckgo-search
markdown==3.8.1
    # via tensorboard
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
matplotlib==3.10.3
    # via coqui-tts
monotonic-alignment-search==0.2.0
    # via coqui-tts
more-itertools==10.7.0
    # via inflect
mpmath==1.3.0
    # via sympy
msgpack==1.1.1
    # via librosa
multidict==6.5.0
    # via
    #   aiohttp
    #   yarl
networkx==3.5
    # via
    #   gruut
    #   torch
num2words==0.5.14
    # via
    #   coqui-tts
    #   gruut
numba==0.61.2
    # via librosa
numpy==2.2.6
    # via
    #   -r requirements.in
    #   contourpy
    #   coqui-tts
    #   coqui-tts-trainer
    #   ctranslate2
    #   encodec
    #   gruut
    #   librosa
    #   matplotlib
    #   monotonic-alignment-search
    #   numba
    #   onnxruntime
    #   scikit-learn
    #   scipy
    #   soundfile
    #   soxr
    #   tensorboard
    #   transformers
ollama==0.5.1
    # via -r requirements.in
onnxruntime==1.22.0
    # via faster-whisper
packaging==25.0
    # via
    #   coqui-tts
    #   coqui-tts-trainer
    #   huggingface-hub
    #   lazy-loader
    #   matplotlib
    #   onnxruntime
    #   pooch
    #   tensorboard
    #   transformers
pillow==11.2.1
    # via
    #   -r requirements.in
    #   matplotlib
platformdirs==4.3.8
    # via pooch
playsound==1.2.2
    # via -r requirements.in
pooch==1.8.2
    # via librosa
primp==0.15.0
    # via duckduckgo-search
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
protobuf==6.31.1
    # via
    #   onnxruntime
    #   tensorboard
psutil==7.0.0
    # via
    #   -r requirements.in
    #   coqui-tts-trainer
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via ollama
pydantic-core==2.33.2
    # via pydantic
pyee==13.0.0
    # via aiortc
pylibsrtp==0.12.0
    # via aiortc
pyopenssl==25.1.0
    # via aiortc
pyparsing==3.2.3
    # via matplotlib
pyreadline3==3.5.4
    # via humanfriendly
pysbd==0.3.4
    # via coqui-tts
python-crfsuite==0.9.11
    # via gruut
python-dateutil==2.9.0.post0
    # via
    #   dateparser
    #   matplotlib
python-dotenv==1.1.0
    # via -r requirements.in
python-engineio==4.12.2
    # via python-socketio
python-socketio==5.13.0
    # via flask-socketio
pytz==2025.2
    # via dateparser
pyyaml==6.0.2
    # via
    #   coqui-tts
    #   ctranslate2
    #   huggingface-hub
    #   transformers
regex==2024.11.6
    # via
    #   dateparser
    #   transformers
requests==2.32.4
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   pooch
    #   transformers
safetensors==0.5.3
    # via transformers
scikit-learn==1.7.0
    # via librosa
scipy==1.15.3
    # via
    #   -r requirements.in
    #   coqui-tts
    #   librosa
    #   scikit-learn
simple-websocket==1.1.0
    # via python-engineio
six==1.17.0
    # via
    #   jsonlines
    #   python-dateutil
    #   tensorboard
sniffio==1.3.1
    # via anyio
sounddevice==0.5.2
    # via -r requirements.in
soundfile==0.13.1
    # via
    #   coqui-tts
    #   coqui-tts-trainer
    #   librosa
soxr==0.5.0.post1
    # via librosa
sympy==1.14.0
    # via
    #   onnxruntime
    #   torch
tensorboard==2.19.0
    # via coqui-tts-trainer
tensorboard-data-server==0.7.2
    # via tensorboard
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.21.1
    # via
    #   faster-whisper
    #   transformers
torch==2.7.1
    # via
    #   coqui-tts
    #   coqui-tts-trainer
    #   encodec
    #   torchaudio
torchaudio==2.7.1
    # via
    #   coqui-tts
    #   encodec
tqdm==4.67.1
    # via
    #   coqui-tts
    #   faster-whisper
    #   huggingface-hub
    #   transformers
transformers==4.51.3
    # via coqui-tts
typeguard==4.4.4
    # via inflect
typing-extensions==4.14.0
    # via
    #   anyio
    #   coqpit-config
    #   coqui-tts
    #   huggingface-hub
    #   librosa
    #   pydantic
    #   pydantic-core
    #   pyee
    #   pyopenssl
    #   torch
    #   typeguard
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via tzlocal
tzlocal==5.3.1
    # via dateparser
urllib3==2.5.0
    # via requests
werkzeug==3.1.3
    # via
    #   flask
    #   tensorboard
wsproto==1.2.0
    # via simple-websocket
yarl==1.20.1
    # via aiohttp
face-recognition==1.3.0
    # via -r requirements.in
face-recognition-models==0.3.0
    # via face_recognition

# The following packages are considered to be unsafe in a requirements file:
# setuptools
